Make specific unit work for combat log event triggers
The text entry for specific unit has a derived its hidden function
incorrectly from the parent's hidden function.

The two changes:
a) Pass in the trigger to the hidden function
b) If hidden was a function the "or hidden" part would always be
   true, since a function is a true value.

Making those options visible seem to be enough to actually fix
the ticket 247.