Model: Reset model on UNIT_MODEL_CHANGED and a few other events
We used to set the model everytime the trigger triggered, but
that lead to frequent stuttering of the model animation.

This patch watches UNIT_MODEL_CHANGED and checks if the event
is for the unitId we are watching. It also wathces
PLAYER_TARGET_CHANGED and PLAYER_FOCUS_CHANGED if the unitId is
target or focus.