Fix editing multiple auras with multiple triggers
Testcase that shows the problems:
Import http://pastebin.com/rS4S5ePq, that is a group with two auras.
Both auras contain 3 identical triggers.

Bug 1:
Select the group, switch to the trigger tab. Notice that only
two triggers are listed in the choose trigger combo box.

Bug 2:
Choose both auras (not the group), switch to the trigger tab,
in the choose trigger combo select the second or third trigger