BuffTrigger: SpellId triggers should match all auras with the same name
That is for non fullscan auras and given a spellId, the buff trigger
should only match the name of the spellId.

This makes some use cases easier, e.g. a cross-language trigger for
a boss ability that has multiple different spell ids.

On the other hand matching on the spellId in addition adds no benefit.

Ticket-Nr: 554