Fix combat log trigger not showing the right duration
The state for generic triggers is updated in two steps. 1) By the trigger
function and 2) by ActivateEvent which adds all the other dynamic
information. If a trigger function is a "all" trigger function,
that is it supports cloning it needs to indicate on which states
ActivateEvent must be called on. Do that by setting changed to true
in those cases.

Ticket-Nr: 564