Options: Ensure that certain ui elements start on a new line
By inserting a empty description before them.

This can happen if a group has two auras:
Aura 1, the spec selection is in single selection mode.
Aura 2, the spec selection is in multi selection mode.

That is if the user selects aura1, the combobox is show,
the group of checkboxes hidden.

If aura2 is selected, the combobox is hidden and the checkboxes
are shown.

If now a group of those auras is selected, the "hiddenAll" function
checks for each widget if the widget is visible for all auras.

=> Both widgets are hidden and no widget is shown in the right column.
Thus the next widget which should be in the left column is shown in the
right.

Ticket-Nr: 562