Remove recent_chages code from BuffTrigger.lua
This caused some auras to not show up on loading.

In the bug reports case, the code was tracking buffs only
outside of combat.

The sequence of events:
- Apply the buff
- Enter combat, unloading the aura
- Leave combat, loading the aura

Loading a aura causes a ScanAll(), in the bug reports case
this would call AssertAura, but due to the data being identical
to the data gathered before combat, the aura_cache would indicate
no change. Yet, after a load we want to update everything.

I think this has been always broken, as I don't recall changing this
logic.

The fix is, ignoring whether the aura_cache already knew that buff.
Since the return value of AssertAura/DeassertAura is the only place
where recent_changes is used, purge it from everywhere.

Ticket-Nr: 645