Fix dynamic groups again by releasing clones at a different place
A bit of background information:
Clones are allocated/released to a clone pool as needed so that
they are reused.

A long time ago, clones were released in a OnHide event handler.
That was wrong, because e.g. movies hide all regions, and there
were thus a few bugs where movies would confuse weakauras.

I fixed that by decoupling hiding and releasing, and made WA
only release the clones if WA is doing the hiding.

In that patch, I misunderstood the interaction with dynamic groups,
and thus clones from dynamic groups were never released.

A recent bug report reported performance problems, which I tracked
down to this bug.

This patch now moves the releasing to the right place.

Ticket-Nr: 674, 708