Sort the event/status/sound/chat dropdown's list
By adding a DropDown that automatically sorts the entries. This control is
just a normal DropDown with a overriden SetList function. The normal
DropDown can already sort, but due to AceConfig always calling SetList
with a order parameter set to nil, it never does sort.

The SortedDropDown automatically sorts. Add it to a few places where
the list is big enough that sorting improves the user experience.