Change many triggers to not automatically hide
If there are two triggers with inverse conditions, e.g. a cooldown
active + cooldown not active trigger. We ideally want to always have
exactly one of those auras active. If we automatically hide the
cooldown active trigger on expiration time, this can be a moment
before getting the actual event.

So for those triggers where we expect a event, disable the automatic
hiding.

Ticket-Nr: 793