Fix ignore rune cd again
Broke that accidentally with the change for Epidemic.

The intent behind this code is this:

Epidemic is a spell with charges, thus GetSpellCharges is the right api.

We compare durations between the cooldown duration and rune cd duration
to figure out whether a spell is on cooldown only for a rune.

But Epidemic has the same cooldown duration as runes.

Luckily, GetSpellCharges doesn't adjust it's output based on rune cds.

Thus, we can be sure that any cooldown returned by it cannot be due
to a rune cd.

Ticket-Nr: 835