Fix TextEditor for group editing if paths differ
The TextEditor gets a "path" to the setting it is modifying, e.g.
data.additional_triggers.3.trigger. For group auras, the function
appendToTriggerPath tried to figure out the common path.

This fails in the linked bug report, because the custom function
is in trigger 2 in one aura and in trigger 3 in a different aura.

Instead, build a "multipath" structure in appendToTriggerPath, that
is a data structure which notes for each id what the path is, and
teach the TextEditor how to handle that.

Ticket-Nr: 959