Change how text replacements work
Instead of trying all dynamic texts and all members of state,
iterate the string once. And on finding a %, use all alphabetical
characters after it. If that string is only one character look it
up via dynamic texts. Otherwise look it up via the state.

Since we are only iterating once, instead of many times this should
also be faster than before.

Note: This is not 100% backwards compatible, e.g. %pRage would
previously replace the %p, this now requires a space.

Ticket-Nr: 957