Fix potential Stack overflow (#217)
As it happens, there actually is a use for calling WeakAuras.ScanEvents inside of the init script of an aura: doing so triggers ActivateAuraEnvironment, which then runs the appropriate init script if it hasn't completed yet; handy if you want to ensure that one aura (a library of sorts) which defines several functions you use elsewhere is loaded first. 

However, the code has no notion of whether or not an init script has begun execution or not, which leads to the possibility of a ScanEvents chain leading back to your original init script. This leads to infinite recursion, and very quickly a stack overflow.