Extract BuffTrigger and GenericTrigger options from WeakAurasOptions
There's still some code in WeakAurasOptions.lua that is specific to the
trigger system used. And the code in WeakAurasOptions very much knows
that there are Buff and Generic triggers.

But this is a step towards making triggersystems register their
options and the WeakAurasOptions code not being specific to any
trigger system.