Fix check in checkConditions for aura visibility
In this issue, a condition is on the remaining time of a cooldown.
If the aura is shown before the remaining time is up, we schedule
a conditionCheck for the future.

That code needs to then check if the aura is still visible before
running the conditions. Previously the code checked
region.state.show, but that's the state of one of the triggers and
not the overall visibility of the aura.

Github-Issue: 332