Fix anchored auras in dynamic groups
ticket # 456 - If a child of a dynamic group has anchorFrameType of SELECT and anchorFrameFrame refers to a frame which doesn't exist, then WeakAuras will get stuck in a loop trying to anchor, which eventually causes the group to fail to control its children. We can prevent this by forcing GetAnchorFrame to detect a dynamicgroup parent.