Fix MoverResizer not getting the right size after import
The MoverResizer depends on the data that is stored in the region's
trx, try, blx, bly members. Those values are set in the modify() of
the group and depend on all children.

Thus, if we change the children of a group, we need to always also
add the parent, so that those values are correct.

Reported on discord.