Fix false negative on sortHybridTable change
Since sortHybridTable goes by child ID, which can be quite volatile,
we need to detect a real change more carefully than WeakAuras.diff can reasonably do.
Also, there was a bug in importPendingData when merging sortHybridTable,
where it would mark a child as part of the hybrid sort if either its new
or old version is part of the hybrid sort, and that is wrong. Now, it properly
merges based on the Group Arrangement user preference.
Also preserveOldOrder variable's name was changed to better reflect what it is now.