add conditions to templates (#587)
* add conditions to templates

* templates can have multiple types
* for every "ability" template, added "abilityShowAlways" or "abilityCharge" template
abilityChargeBuff and abilityChargeDebuff for spell with charge + show buff/debuff when active
abilityChargeTarget for spell with charge + rangecheck
* icon templates show cooldown by default (was disabled for auras)
"Has Target" condition added to "debuffShowAlways"
added "debuffShowAlways" to every "debuff" template, same for buffs and items
* abilityTotem & abilityChargeTotem templates
* Change how conditions and trigger variants are built
The basic concept is, the database describes what a spell is. There
are a few new flags for that:
- charges: If the spell has charges
- usable: If a usable check can be useful
- requiresTarget: If the spell requires a target
- buff: If the ability buffs the player
- debuff: If the ability debuffs the target
- totem: If the ability places down a totem

The third trigger page now selects the "subtype". A subtype can decide
how exactly the triggers and conditions are setup. A subtype has
two functions: createTriggers and createConditions which do that, and
builts its conditions from a list of common changes.

The mapping between spell properties and subtypes is done via
subTypesFor. (Which can properly be refactored a bit.)

* use insufficientResourcesBlue for abilities without charges=true
use spellInRangeRed for abilities with requiresTarget

* added optional totemType option, usefull when totemName doesn't match spellId
updated monk's refreshing jade wind, black ox statue and jade serpent statue