Create an Aura Environment for group regions as well.
This is necessary because the Animated Expand & Collapse feature animates the dynamic group region itself.
This eventually calls ActivateAuraEnvironment, which tries to get the aura environment for the dynamic group.
Before, this would fallback to an empty table, which is rather wasteful since a new extra table would be
created every frame for every dynamic group which is animating an expand or collapse.