Fix imports of pre version 7 auras
We used to minimize auras by removing defaults, That requires that
defaults don't change, which is not the case. Thus this is a dumb idea.

So remove it. There are lots of auras without defaults already
exported, so for those readd the right defaults depending on
the internal version.

This also means that the validate can't happen before Modernize.

Github-Issue: 733