Workaround another Blizzard bug with text handling
Or maybe it's the same bug all along.
If we change the font in configureText, we need to redo our text
calculation and set the "right" width.

Somehow this would not correctly justify the text, so the text wouldn't
appear in the bottom right/bottom left corner but would have a
horizontal offset. (Could be reproduced by switching to a different text
positon with the options open.)

Setting the text temporarily to a different value and back to the
original, seems to unlock something in Blizzard's code. It probably
leads to some kind of lazy layout flag being set.

After doing that SetJustifyH works together with the width setting.

Ticket-Nr: 1220