Send ENCOUNTER_START event triggers with a encounter load option
Before 0bfb77d the generic trigger system called RegisterEvent on all
the events for all the auras, even if they weren't loaded.

In 0bfb77d this was changed to be done on aura loading instead. This
means that the generic trigger system is not registered for
ENCOUNTER_START until a aura which registers that event is loaded.

To fix that, forward the event that leads to the loading of auras to the
trigger systems. (And since multiple auras can load, do a batch
processing of auras.)

Thus the generic trigger system can recognize that the event is a event
that it will registered by loading those auras and was missed due to the
on demand registration and handle the event properly.

Github-Issue: 262