Fix SetRegion hiding groups
In 1779683 SetRegion was changed to hide all regions initially.

While auras are shown/hidden after processing a trigger update, groups
are unaffected by that and need to be always visible.

Argueably they should the check the status of their children, but that's
bigger change, so just revert it for groups/dynamicgroups.