Icon: Make "Show Cooldown Text" toggle work with OmniCC
Turns out that there's indeed a way to tell OmniCC that information,
as pointed via the ticket.

Also rename and inverse the option to "Hide Cooldown Text" and make it
always visible.

We used to hide the option also for the case where the user was using
Blizzard's cooldowns but with the text option not enabled. In that case
the option obviously had no effect.

But, showing the option allows aura authors to use it to disable the
cooldown text for all users.

For the same setup - Blizzard without Text -, having the option be
"Show Cooldown Text" would have been odd, as enabling it would not have
shown any text. The renaming fixes that.

Ticket-Nr: 1320