Cooldown Progress/Action Usable: Change how spell count is used
This breaks some auras, but after some consideration this is needed
sooner or later. The breaking change is that %charges and the
trigger/condition check don't consider spell counts anymore.

%s remains magically: Show charges if the spell has charges or spell
counts if the spell count is > 0. This should match what is displayed on
action buttons.

%charges + trigger check + condition uses GetSpellCharges or if the
spell has no charges pretends that a not on cooldown spell has 1 charge,
and otherwise 0 charges.
This is to e.g. make lava eruption charges checks work consistently
regardless on whether the player has the echo talent.

%spellCount + trigger check + condition is the new spell count only
number. This is to be used with the few spells that are using spell
counts such as:
* Expel Harm
* Marrowrend
* Spinning Crane Kick

Fixes: #1321