Generic Trigger: Fix specific unit not showing up in some cases
For a specific unit, there are two settings that need to be stored,
the flag that the trigger is in specific unit mode, and the specific
unit's name.

The bufftrigger quite sanely does that via:
 flag: unit == "member"
 name: specific_unit

The generic trigger does it via:
 flag: use_specific_unit == true
 name: unit

 and if the user selects specific unit in the combobox, the setter sets
 the flag.

This rather stupid, but due to the trigger code then all depending on data.unit
being either the unit, or the specific unit's name, this can't be changed
without major changes. Thus work around it.

Fixes: #1433