Fix conditions not untriggering in certain conditions
The bug was a aura with:
* a desaturate condition if spell on cd
* a fade out animation

On closing the options.

In that case, the desaturate condition is now active in the options due
to the fake progress we display.

On closing the options, we first
- collapse all regions. Since the aura has a finish animation, the
conditions aren't unapplied immediately, but on finishing the animation.
- and then unload the aura
- then load the aura, which clears the table that stores which
conditions are currently active.

Thus, the currently active table and the actual state didn't agree
anymore.

Fix that by canceling the finish animations on unloading.

Fixes: #1720