Add VEHICLE_UPDATE to zone load options
In a recent commit I optimized our load handling, so that which auras
load state is checked depens on which load options they have.

Unfortunately the Ashenvane -> Orgonzoa Teleport does not send any
zone related event. Previously this worked via the VEHICLE_UPDATE
event.

So as a workaround add that event to the zone related load options.

Fixes: #1695