Fix nil error on deleting a trigger
This is unfortunate. In pAdd we want to do the following danace:

- Collapse all clones
- Add the new data to db.displays
- Show new clones based on the now set data

Unfortunately, even before pAdd sets the data in db.displays,
db.display, db.displays already contains the new data as we share the
same data table everywhere.

And that leads to a error in collapse, which expects there to be a
db.displays[id].triggers[triggernum] for a trigger that was just
deleted.

The right fix would be to make sure that db.displays can only be
modified via pAdd, but that's infeasiable.

Fixes: #1886