DynamicGroups: Fix text placement not adjusting to text dimensions
Texts are special in that they don't have a hardcoded size but instead
dynamicyally adjust their size based on contents.

And dynamic groups have supported that special case by using
region.width instead of data.width.

Thus the code in dynamic groups was e.g. for width
'data.width or region.width'.

That has the nice property of not hardcoding the specialness.

But if the user converts a existing region to a text, it has a width
setting in data.

Thus make the dynamic group aware that texts are special. Make this a
bit nicer by deciding what to use in exactly one place though.

Fixes: #1999