Fix specific unit for unit type triggers
After some testing reveals that RegisterUnitEvent with e.g .
targettarget "works" in a limited way. Unfortunately we did
accidentally worked in that limited way before the unit trigger
rewrite.

Specifally, RegisterUnitEvevent for a non base unitid, e.g. targettarget
will in general not deliver events. Except if the targettarget unit is
known under a base unit id, then events will be delivered. E.g. if
targettarget is a raid member or a nameplate. Those events will have the
first argument set to the base unit id that provides the data.

Thus: Register even non base unitids, and on getting events deliver them
with the unit id that we registered, instead of the unit id of the
event.

Fixes: #2148