Fix anchoring to unit frames on group join
On joing a group the frame for a unit might not yet be available.
LibGetFrame informs us that a new association between unit and frame was
found.

In that case, we need to more than just reanchor a individual region, we
need to reevaluate the position for all children.

Calling DoPositionChildren is a pretty big hammer, but this works for
now.

Fixes: #2480