Fix the merged zone id load option
Nightwarden pointed out that there are duplicated is, and thus the
merging was actual a mistake.

Recover from that mistake by adopting a convention that group zone ids
must be prefixed with 'g'. Obviously that is a bit strange, translates
badly, and is a entirely new concept that we don't use anywhere else.

But, it keeps the train moving forward...