Make BuffTrigger 2 hide the per match checks if not applicable
That is if a trigger is set in this mode:

showOn: Always
showClones: false

Previously options for e.g. a remaining time, a stacks check and similar
would be shown.

They worked in a somewhat suprising way, since always gurantees that if
no state matches a empty state is shown.

It's probably better to not enable these options at all.