Options: Fix moving auras showing wrong text replacements
The cause of this was df9140cbaa, which fixed a bug where states were
almost always applied to regions.

That exposed a bug, that after moving a aura around, the region's modify
function resets the text, which then isn't overwritten by a state.

Because the ApplyStatesToRegion correctly thinks that the region is
already synced with the state.

So reset state in modify, because fundamentally modify breaks any sync
with a state.

Fixes: #3536