Remove data.controlledChildren from auras that aren't groups
This is the cause behind Tickets: #3964 and #3957, in both cases there
was an aura in the data that is not a group, but had a
controlledChildren member

The former had this aura: https://wago.io/r1JHvtSz7, which was
in previous versions a group aura, and then later switched to
be a single progress texture.

The later had thsi aura: https://wago.io/HB29dQYlW/16, which similarly
used to be a group aura, before switching to a text aura.

The current upgrade code relies on uid matching and explicitly does not
support uprading if a aura switched from group type to non-group type.

Thus my most likely explanation is that the old upgrade code had a bug
when upgrading a group aura to a non-group, which left alone the
controlledChildren member. But that this bug didn't cause any issues
until recently.

Fixes: #3964
Fixes: #3957