Fix multi-selection with new trigger/load layout
Due to how the multi-selection works we need to always create the dummy
spacer item and the real controls and hide the ones that we don't need.

But the default hiddenAll doesn't work for the dummy widget, as we only
want it to show if the input widget is hidden, thus add another hack to
our options code.

Nothing bad can come from that.

Fixes: #4604