only use whisper comms if target is on the actually same server (#5019)
* only use whisper comms if target is on the actually same server

UnitIsSameServer returns true if user is on a connected realm (e.g. US ArgentDawn <=> TheScryers), so we should distinguish between the cases using UnitRealmRelationship, which returns 1 for same realm, 2 for coalesced (instanced/sharded together or whatever), and 3 for connected realms.

Since it seems that either whisper comms on connected realms never worked (and ~nobody plays on those realms so it took a long time to notice), or blizzard recently broke it, we simply fallback to trying raid/party. If that doesn't work, we already have a hint to the user to try grouping up.

Signed-off-by: Allen Faure <afaure6@gmail.com>