Workaround INSTANCE_ENGAGE_UNIT with incosistent UnitGUID/UnitExists
While testing the first boss of Underrot sometimes the game fired two
INSTANCE_ENGAGE_UNIT events for one additional clone.

On the first INSTANCE_ENGAGE_UNIT, UnitGUID returns the correct guid,
but other apis such as UnitExists/UnitHealth etc treat the unit as not
existing.

On the second INSTANCE_ENGAGE_UNIT, the unit fully exists. But we only
fired a UNIT_CHANGED_ event for the first event. Fix that by tracking
the UnitExists return value and also firing UNIT_CHANGED_ if that
changes.